+++ src/components/$.ts
@@ -1,6 +1,6 @@
 import { Kind } from "hkt-toolbelt";
 
-export type $<F extends Kind, X extends Kind.Input<F>> = ReturnType<
+export type $<F extends Kind, X extends Kind.InputOf<F>> = ReturnType<
   (F & {
     readonly [Kind._]: X;
   })["f"]