+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.1
+
+- Changed `Kind.Input` to `Kind.InputOf`
+- Changed `Kind.Output` to `Kind.OutputOf`
+  - These were changed to improve clarity.
+
 ## 0.2.0
 
 - Added HKT-level composability check for `Kind.Compose`.